Update README.md
The game Floaty Kite is a simple play on the popular game Flappy Bird, but, well, with a Kite instead of a bird. This assignment was for fun, and for my final assignment in my Object Oriented Programming class at Arizona State. There is no, or never will there be, intent on using this game as a means of generating a profit. The future of this project may include a means of saving an all time high score instead of resetting he high score on exit, the inclusion of Nintendo like pipes and clouds, and the inclusion of sound.